Music functionality is at 90% Only missing Shuffle & Loop slash commands for music also implemented
